Refactor reducers to extra files
[MAILPOET-2450]
