+++ assets/js/src/form_editor/store/reducers/changeFormName.jsx
@@ -1,0 +1,7 @@
+export default (state, action) => ({
+  ...state,
+  formData: {
+    ...state.formData,
+    name: action.name,
+  },
+});
