+++ assets/js/src/form_editor/store/reducers/removeNotice.jsx
@@ -1,0 +1,4 @@
+export default (state, action) => ({
+  ...state,
+  notices: [...state.notices].filter((item) => item.id !== action.id),
+});
