+++ assets/js/src/form_editor/store/reducers/saveFormDone.jsx
@@ -1,0 +1,4 @@
+export default (state) => ({
+  ...state,
+  isFormSaving: false,
+});
