+++ assets/js/src/form_editor/store/reducers/toggleSidebar.jsx
@@ -1,0 +1,4 @@
+export default (state, action) => ({
+  ...state,
+  sidebarOpened: action.toggleTo,
+});
